safari-600.7-branch/LayoutTests/fast/html/clone-keygen-expected.txt

Tests that a cloned keygen has the right shadow pseudo ID. You should see two big keygens with purple text (they should look the same.)


PASS specialK.clientHeight is k.clientHeight
PASS internals.shadowPseudoId(cloneShadow.firstChild) is internals.shadowPseudoId(kShadow.firstChild)
PASS successfullyParsed is true

TEST COMPLETE
